adding IMissingDependencyExpressionProvider for issue #169
