+++ src/Grace/DependencyInjection/Impl/ExportRegistrationBlock.cs
@@ -17,6 +17,7 @@ namespace Grace.DependencyInjection.Impl
         private ImmutableLinkedList<IActivationStrategyInspector> _inspectors = ImmutableLinkedList<IActivationStrategyInspector>.Empty;
         private ImmutableLinkedList<IInjectionValueProvider> _valueProviders = ImmutableLinkedList<IInjectionValueProvider>.Empty;
         private ImmutableLinkedList<IMissingExportStrategyProvider> _missingExportStrategyProviders = ImmutableLinkedList<IMissingExportStrategyProvider>.Empty;
+        private ImmutableLinkedList<IMissingDependencyExpressionProvider> _missingDependencyExpressionProviders = ImmutableLinkedList<IMissingDependencyExpressionProvider>.Empty;
         private ImmutableLinkedList<IMemberInjectionSelector> _globalSelectors = ImmutableLinkedList<IMemberInjectionSelector>.Empty;
         private readonly IActivationStrategyCreator _strategyCreator;
         private IExportStrategyProvider _currentProvider;
@@ -104,6 +105,15 @@ namespace Grace.DependencyInjection.Impl
         public IEnumerable<IMissingExportStrategyProvider> GetMissingExportStrategyProviders()
         {
             return _missingExportStrategyProviders;
+        }
+
+        /// <summary>
+        /// Get list of missing dependency expression provider
+        /// </summary>
+        /// <returns></returns>
+        public IEnumerable<IMissingDependencyExpressionProvider> GetMissingDependencyExpressionProviders()
+        {
+            return _missingDependencyExpressionProviders;
         }
 
         /// <summary>
@@ -579,6 +589,15 @@ namespace Grace.DependencyInjection.Impl
         }
 
         /// <summary>
+        /// Add missing dependency expression provider
+        /// </summary>
+        /// <param name="provider"></param>
+        public void AddMissingDependencyExpressionProvider(IMissingDependencyExpressionProvider provider)
+        {
+            _missingDependencyExpressionProviders = _missingDependencyExpressionProviders.Add(provider);
+        }
+
+        /// <summary>
         /// Add your own custom activation strategy
         /// </summary>
         /// <param name="activationStrategy">activation strategy</param>
@@ -633,5 +652,6 @@ namespace Grace.DependencyInjection.Impl
                 _currentProvider = null;
             }
         }
+
     }
 }
